Separated OCR from pipeline proper
