+++ papermerge/test/test_import_pipelines.py
@@ -3,7 +3,6 @@ import string
 import os
 
 from django.test import TestCase, override_settings
-from mglib.step import Step
 
 from .utils import create_root_user
 from papermerge.core.import_pipeline import (
@@ -13,7 +12,6 @@ from papermerge.core.import_pipeline import (
     WEB,
     DefaultPipeline
 )
-from papermerge.core.storage import default_storage
 
 BASE_DIR = os.path.dirname(__file__)
 
@@ -143,111 +141,6 @@ class TestDefaultPipelineBytes(TestCase):
             apply_kwargs = self.make_apply_kwargs()
             doc = go_through_pipelines(init_kwargs, apply_kwargs)
             self.assertIsNotNone(doc)
-
-
-class TestOCR(TestCase):
-    def setUp(self):
-        self.user = create_root_user()
-
-    def make_init_kwargs(self, payload, processor):
-        return {'payload': payload, 'processor': processor}
-
-    def make_apply_kwargs(self, apply_async=False):
-        return {'skip_ocr': False, 'apply_async': apply_async}
-
-    @override_settings(PAPERMERGE_PIPELINES=PAPERMERGE_DEFAULT_PIPELINE)
-    def test_default_pipeline_jpg_ocr_async(self):
-        file_path = os.path.join(
-            BASE_DIR,
-            "data",
-            "page-1.jpg"
-        )
-        with open(file_path, 'rb') as f:
-            payload = f.read()
-        init_kwargs = self.make_init_kwargs(
-            payload=payload, processor='TEST')
-        apply_kwargs = self.make_apply_kwargs(apply_async=True)
-        doc = go_through_pipelines(init_kwargs, apply_kwargs)
-        self.assertIsNotNone(doc)
-        page_path = doc.get_page_path(
-            page_num=1,
-            step=Step(0),
-        )
-        img_abs_path = default_storage.abspath(
-            page_path.img_url()
-        )
-        self.assertTrue(os.path.exists(img_abs_path))
-
-    @override_settings(PAPERMERGE_PIPELINES=PAPERMERGE_SIMPLE_PIPELINE)
-    def test_simple_pipeline_jpg_ocr_async(self):
-        file_path = os.path.join(
-            BASE_DIR,
-            "data",
-            "page-1.jpg"
-        )
-        with open(file_path, 'rb') as f:
-            payload = f.read()
-        init_kwargs = self.make_init_kwargs(
-            payload=payload, processor='TEST')
-        apply_kwargs = self.make_apply_kwargs(apply_async=True)
-        doc = go_through_pipelines(init_kwargs, apply_kwargs)
-        self.assertIsNotNone(doc)
-        page_path = doc.get_page_path(
-            page_num=1,
-            step=Step(0),
-        )
-        img_abs_path = default_storage.abspath(
-            page_path.img_url()
-        )
-        self.assertTrue(os.path.exists(img_abs_path))
-        self.assertEqual(doc.name, 'test_change_name')
-
-    @override_settings(PAPERMERGE_PIPELINES=PAPERMERGE_DEFAULT_PIPELINE)
-    def test_default_pipeline_jpg_ocr_noasync(self):
-        file_path = os.path.join(
-            BASE_DIR,
-            "data",
-            "page-1.jpg"
-        )
-        with open(file_path, 'rb') as f:
-            payload = f.read()
-        init_kwargs = self.make_init_kwargs(
-            payload=payload, processor='TEST')
-        apply_kwargs = self.make_apply_kwargs()
-        doc = go_through_pipelines(init_kwargs, apply_kwargs)
-        self.assertIsNotNone(doc)
-        page_path = doc.get_page_path(
-            page_num=1,
-            step=Step(0),
-        )
-        img_abs_path = default_storage.abspath(
-            page_path.img_url()
-        )
-        self.assertTrue(os.path.exists(img_abs_path))
-
-    @override_settings(PAPERMERGE_PIPELINES=PAPERMERGE_SIMPLE_PIPELINE)
-    def test_simple_pipeline_jpg_ocr_noasync(self):
-        file_path = os.path.join(
-            BASE_DIR,
-            "data",
-            "page-1.jpg"
-        )
-        with open(file_path, 'rb') as f:
-            payload = f.read()
-        init_kwargs = self.make_init_kwargs(
-            payload=payload, processor='TEST')
-        apply_kwargs = self.make_apply_kwargs()
-        doc = go_through_pipelines(init_kwargs, apply_kwargs)
-        self.assertIsNotNone(doc)
-        page_path = doc.get_page_path(
-            page_num=1,
-            step=Step(0),
-        )
-        img_abs_path = default_storage.abspath(
-            page_path.img_url()
-        )
-        self.assertTrue(os.path.exists(img_abs_path))
-        self.assertEqual(doc.name, 'test_change_name')
 
 
 class PipelineOne(DefaultPipeline):
